requirements.txt
MetaTrader5>=5.0.45
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
